backend/requirements.txt
fastapi>=0.100.0
pydantic>=2.0.0
uvicorn>=0.22.0
python-multipart>=0.0.6
pillow>=9.5.0
httpx>=0.24.0
azure-storage-blob>=12.0.0